+++ apps/src/code-studio/pd/workshop_dashboard/workshops/surveys/post/components/FacilitatorFeedback.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const FacilitatorFeedback = () => (
+  <div>Individual Facilitator Feedback</div>
+);
